# api/openapi.yaml
openapi: 3.1.0
info:
  title: 食堂窗口叫号 API
  version: 0.1.0
  description: HTTP 合同权威。错误码与 DTO 只在此文件与 SPEC 出现一次。
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      summary: 学号密码登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              student_id: "20260001"
              password: "diner123"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: 学号或密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/logout:
    post:
      summary: 退出登录并使当前令牌失效
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 已退出
        "401":
          description: 未登录
  /me:
    get:
      summary: 当前会话用户
      operationId: me
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: 未登录
  /admin/ping:
    get:
      summary: 管理员探测（用于角色隔离验收）
      operationId: adminPing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 管理员可访问
        "403":
          description: 非管理员
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /canteens:
    get:
      summary: 食堂列表
      operationId: listCanteens
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 食堂列表
        "401":
          description: 未登录
  /canteens/{id}/floors:
    get:
      summary: 食堂楼层列表
      operationId: listFloors
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 楼层列表，无楼层时 items 为空
        "401":
          description: 未登录
        "404":
          description: 食堂不存在
  /windows:
    get:
      summary: 窗口摘要列表
      operationId: listWindows
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: canteen_id
          required: true
          schema:
            type: string
        - in: query
          name: floor_id
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 窗口摘要，含 current_number 与 waiting_count
        "401":
          description: 未登录
  /tickets:
    post:
      summary: 取号（同一窗口幂等返回已有号）
      operationId: takeTicket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [window_id]
              properties:
                window_id:
                  type: string
      responses:
        "200":
          description: 票对象
        "401":
          description: 未登录
        "409":
          description: 已有其它窗口有效号，或窗口不可取号
  /me/ticket:
    get:
      summary: 我的有效号
      operationId: myTicket
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 有效票或 data=null
        "401":
          description: 未登录
  /tickets/{id}/cancel:
    post:
      summary: 取消等待中的号
      operationId: cancelTicket
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 已取消的票
        "401":
          description: 未登录
        "409":
          description: 不是等待中的号
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Envelope:
      type: object
      required: [ok, data, error]
      properties:
        ok:
          type: boolean
        data: {}
        error:
          oneOf:
            - $ref: "#/components/schemas/APIError"
            - type: "null"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - VALIDATION_ERROR
            - WINDOW_NOT_OPEN
            - WINDOW_PAUSE_TAKE
            - HAS_ACTIVE_TICKET
            - TICKET_NOT_WAITING
            - TICKET_NOT_CALLED
            - NO_WAITING
            - CALLED_PENDING
            - NOT_FOUND
            - CONFLICT
            - DISPLAY_TOKEN_INVALID
            - INTERNAL_ERROR
        message:
          type: string
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    LoginRequest:
      type: object
      required: [student_id, password]
      properties:
        student_id:
          type: string
        password:
          type: string
    PublicUser:
      type: object
      required: [id, student_id, role]
      properties:
        id:
          type: string
        student_id:
          type: string
        role:
          type: string
          enum: [diner, staff, admin]
    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: object
          required: [token, user]
          properties:
            token:
              type: string
            user:
              $ref: "#/components/schemas/PublicUser"
        error:
          type: "null"
    MeResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          $ref: "#/components/schemas/PublicUser"
        error:
          type: "null"